feat(#51): moved schedule details to another controller
